requirements.txt
absl-py==0.14.1

asgiref==3.4.1; python_version >= '3.6'

attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

certifi==2021.10.8

charset-normalizer==2.0.7; python_version >= '3'

coreapi==2.3.3

coreschema==0.0.4

cycler==0.10.0

dj-database-url==0.5.0

django-debug-toolbar==3.2.2

django==3.2.8

djangorestframework==3.12.4; python_version >= '3.5'

drf-yasg==1.20.0

gunicorn==20.1.0

idna==3.2; python_version >= '3'

inflection==0.5.1; python_version >= '3.5'

itypes==1.2.0

jinja2==3.0.2; python_version >= '3.6'

kiwisolver==1.3.2; python_version >= '3.7'

markupsafe==2.0.1; python_version >= '3.6'

matplotlib==3.4.3; python_version >= '3.7'

mediapipe==0.8.8

numpy==1.21.2; python_version < '3.11' and python_version >= '3.7'

opencv-python-headless==4.5.3.56

packaging==21.0; python_version >= '3.6'

pillow==8.3.2; python_version >= '3.6'

protobuf==3.18.1; python_version >= '3.5'

psycopg2==2.9.1

pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

python-decouple==3.5

pytz==2021.3

requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'

ruamel.yaml.clib==0.2.6; python_version < '3.10' and platform_python_implementation == 'CPython'

ruamel.yaml==0.17.16; python_version >= '3'

six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

sqlparse==0.4.2; python_version >= '3.5'

torch==1.9.1

torchvision==0.10.1

typing-extensions==3.10.0.2

uritemplate==4.0.0; python_version >= '3.6'

urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

wheel==0.37.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

whitenoise==5.3.0

